+++ configs/general.js
@@ -1,0 +1,5 @@
+module.exports = {
+    "endpoint": {
+        "request_timeout": 1500
+    }
+};