+++ configs/lan.js
@@ -1,6 +1,6 @@
 module.exports = {
-    network_address: "192.168.0.1",
+    network_address: "192.168.137.1",
     mask_bits: 24,
-    default_gateway: "192.168.0.1",
-    scan_query: "192.168.0.0/24"
+    default_gateway: "192.168.137.1",
+    scan_query: "192.168.137.0/24"
 };